Add Export to HTML button in the app GUI (closes #42)
- Extract shared export helpers (build_app_data, embed_photos,
  build_frontend, inline_assets) from export.py into backend/exporter.py
- Add GET /api/export/pick-directory — native macOS folder picker via osascript
- Add GET /api/export?output_dir=... — builds and writes the HTML export
- Add Export to HTML button in the sidebar alongside Sync Contacts
- export.py CLI continues to work unchanged via the shared exporter module
